requirements.txt
# Core dependencies
Flask==3.0.1
gunicorn==21.2.0

# Data processing and ML
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0
joblib==1.3.2

# URL and web processing
requests==2.31.0
beautifulsoup4==4.12.2
python-whois==0.8.0
tldextract==5.1.1
urllib3==2.1.0

# Feature engineering
feature-engine==1.6.1

# Security
python-dotenv==1.0.1
validators==0.22.0
